+++ tests/performance/performance_utils.h
@@ -10,6 +10,9 @@
 #include <iomanip>
 
 #ifdef _WIN32
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
 #include <windows.h>
 #include <psapi.h>
 #pragma comment(lib, "psapi.lib")